# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  post.CreatePostRequest:
    properties:
      body:
        type: string
      category_id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  post.DeletePostResponse:
    properties:
      message:
        type: string
    type: object
  post.GetAllPostsResponse:
    properties:
      posts:
        items:
          $ref: '#/definitions/post.Post'
        type: array
    type: object
  post.Post:
    properties:
      body:
        type: string
      category_id:
        description: UUID
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        description: UUID
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        description: UUID
        type: string
    type: object
  post.UpdatePostRequest:
    properties:
      body:
        type: string
      category_id:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  posttag.CreatePostTagRequest:
    properties:
      post_id:
        type: string
      tag_id:
        type: string
    type: object
  posttag.DeletePostTagRequest:
    properties:
      post_id:
        type: string
      tag_id:
        type: string
    type: object
  posttag.DeletePostTagResponse:
    properties:
      message:
        type: string
    type: object
  posttag.GetAllPostTagsResponse:
    properties:
      post_tags:
        items:
          $ref: '#/definitions/posttag.PostTag'
        type: array
    type: object
  posttag.GetPostsByTagResponse:
    properties:
      posts:
        description: Array of post
        items:
          $ref: '#/definitions/post.Post'
        type: array
    type: object
  posttag.PostTag:
    properties:
      created_at:
        type: string
      post_id:
        description: UUID
        type: string
      tag_id:
        description: UUID
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a forum API gateway service.
  title: Forum API Gateway
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all posts with optional filtering and pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of posts per page
        in: query
        name: limit
        type: integer
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by title
        in: query
        name: title
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: string
      - description: Filter by body content (partial match)
        in: query
        name: body
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.GetAllPostsResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all posts
      tags:
      - post
    post:
      consumes:
      - application/json
      description: Create a new post with given information
      parameters:
      - description: Post information
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/post.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/post.Post'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - post
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a post by its unique ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.DeletePostResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Post item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a post by its ID
      tags:
      - post
    get:
      consumes:
      - application/json
      description: Retrieve a post by its unique ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.Post'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Post item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a post by its ID
      tags:
      - post
    put:
      consumes:
      - application/json
      description: Update a post with the given information
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post information
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/post.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.Post'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Post item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a post by its ID
      tags:
      - post
  /posttags:
    delete:
      consumes:
      - application/json
      description: Delete a post-tag relationship with given information
      parameters:
      - description: PostTag information
        in: body
        name: posttag
        required: true
        schema:
          $ref: '#/definitions/posttag.DeletePostTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posttag.DeletePostTagResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Post_tag item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a post-tag relationship
      tags:
      - posttag
    get:
      consumes:
      - application/json
      description: Retrieve a list of all post-tag relationships with optional filtering
        and pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of post-tag relationships per page
        in: query
        name: limit
        type: integer
      - description: Filter by post ID
        in: query
        name: post_id
        type: string
      - description: Filter by tag ID
        in: query
        name: tag_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posttag.GetAllPostTagsResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all post-tag relationships
      tags:
      - posttag
    post:
      consumes:
      - application/json
      description: Create a new post-tag relationship with given information
      parameters:
      - description: PostTag information
        in: body
        name: posttag
        required: true
        schema:
          $ref: '#/definitions/posttag.CreatePostTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/posttag.PostTag'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new post-tag relationship
      tags:
      - posttag
  /posttags/{tag_id}/posts:
    get:
      consumes:
      - application/json
      description: Retrieve a list of posts associated with a given tag ID
      parameters:
      - description: Tag ID
        in: path
        name: tag_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of posts per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posttag.GetPostsByTagResponse'
        "404":
          description: Posttag item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get posts by tag ID
      tags:
      - posttag
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"